refactor: split dag cascade/status modules, fold OTLP config, and retune structure-reorg rules
