+++ .github/skills/review/references/00-structure-placement.md
@@ -51,7 +51,7 @@ gokit is a multi-module monorepo split by dependency weight and role:
   and to `domains.toml` + the matching `make check-<domain>` set. Missing any is a should-fix.
 - **doc.go present.** Every package has a `doc.go` with a package doc comment. Missing is a should-fix.
 - **Declare-only aggregator.** `doc.go` carries package documentation only — no `func`/`type`/`var`/`const`. Code in a `doc.go`, or a package whose logic is piled into one oversized file instead of concern-named siblings, is a should-fix. Run `scripts/check-structure.sh` (`make structure`).
-- **Package cohesion & file-level structuring.** A package whose non-test code is piled into one **oversized or concern-mixed file** should be split into concern-named sibling files (`client.go`, `options.go`, `types.go`, `errors.go`, `doc.go`), or a large area promoted to a sub-package. Reorg is **not only** about one oversized file: when a single package/folder has accumulated many loosely-related non-test files (roughly >5–10, excluding `_test.go`) that fall into groups which are **not tightly coupled**, each cohesive group should move into its own concern-named sub-package (sub-folder with its own declare-only `doc.go`), so the reader lands in the right file *and folder* from the layout (as in `agent/{memory}`, `cli/{theme,render,…}`). This is **criteria-driven, not a line count**: reorganize only where the split genuinely improves discoverability and maintainability, or where a package already holds many organizable siblings. A cohesive, single-concern file (or a flat set of files that all serve one concern) is fine at any length — do not split for the sake of it. When a change touches such a file, restructure it **in the same change** rather than deferring the reorg (deferred reorg is how structure debt accumulates). The god-file guard (`scripts/check-structure.sh`, `GOD_FILE_LINES` default 600) surfaces candidates; an oversized, concern-mixed file left un-split is a should-fix.
+- **Package cohesion & file-level structuring.** A package whose non-test code is piled into one **oversized or concern-mixed file** should be split into concern-named sibling files (`client.go`, `options.go`, `types.go`, `errors.go`, `doc.go`), or a large area promoted to a sub-package. Reorg is **not only** about one oversized file: when a single package/folder has accumulated **more than ~10 non-test files** (excluding `_test.go`) that fall into **2–3+ separable concern groups**, each cohesive group should move into its own concern-named sub-package (sub-folder with its own declare-only `doc.go`), so the reader lands in the right file *and folder* from the layout (as in `agent/{memory}`, `cli/{theme,render,…}`). This is **criteria-driven, not a line count**: reorganize only where the split genuinely improves readability, maintainability, scalability, extendability, and reusability without causing other issues (import cycles, over-fragmentation), or where a package already holds many organizable siblings. A cohesive, single-concern file (or a flat set of files that all serve one concern) is fine at any length — length alone is never the verdict, concern-mixing is; do not split for the sake of it. When a change touches such a file, restructure it **in the same change** rather than deferring the reorg (deferred reorg is how structure debt accumulates). The large-file guard (`scripts/check-structure.sh`, `GOD_FILE_LINES` default 350, counting code lines only) surfaces candidates; an oversized, concern-mixed file left un-split is a should-fix. All of `make structure` is advisory (never gating).
 - **No misplaced concerns.** Each cross-cutting concern stays in its canonical package —
   e.g. gRPC status mapping belongs in `grpc`, not `errors`. (Reuse of those owners is pass `01`.)
 - **Backend opt-in.** A nested adapter registers via an explicit `Register(registry)` call,
