+++ dag/cascade/doc.go
@@ -1,0 +1,2 @@
+// Package cascade provides staged DAG-style execution where each stage can build provider-backed nodes, order them by metadata, and control stage/final-stage failure behavior.
+package cascade
