+++ dag/result.go
@@ -1,16 +1,9 @@
 package dag
 
-import "time"
+import (
+	"time"
 
-// Node execution status constants.
-const (
-	StatusCompleted      = "completed"               // Ran successfully
-	StatusFailed         = "failed"                  // Ran and returned an error
-	StatusSkipped        = "skipped"                 // Filtered out by schedule/condition
-	StatusUnavailable    = "unavailable"             // Node itself is unavailable (optional + not registered)
-	StatusDepUnavailable = "skipped:dep_unavailable" // Skipped because an upstream was unavailable
-	StatusDepFailed      = "skipped:dep_failed"      // Skipped because an upstream failed
-	StatusDepSkipped     = "skipped:dep_not_ready"   // Skipped because an upstream was skipped and has no state output yet
+	"github.com/kbukum/gokit/dag/status"
 )
 
 // Result holds the outcome of a graph execution.
@@ -22,7 +15,7 @@ type Result struct {
 // NodeResult holds the outcome of a single node execution.
 type NodeResult struct {
 	Name     string
-	Status   string
+	Status   status.Status
 	Duration time.Duration
 	Output   any
 	Error    error
@@ -30,15 +23,15 @@ type NodeResult struct {
 
 // IsTerminal returns true if the node ran (completed or failed).
 func (r NodeResult) IsTerminal() bool {
-	return r.Status == StatusCompleted || r.Status == StatusFailed
+	return r.Status.IsTerminal()
 }
 
 // IsSkipped returns true if the node was skipped for any reason.
 func (r NodeResult) IsSkipped() bool {
-	return r.Status == StatusSkipped || r.Status == StatusDepUnavailable || r.Status == StatusDepFailed || r.Status == StatusDepSkipped
+	return r.Status.IsSkipped()
 }
 
 // IsSuccess returns true if the node completed successfully.
 func (r NodeResult) IsSuccess() bool {
-	return r.Status == StatusCompleted
+	return r.Status.IsSuccess()
 }
