+++ dag/status/doc.go
@@ -1,0 +1,2 @@
+// Package status defines shared DAG execution status values for root DAG execution and DAG subpackages.
+package status
